+++ Editor/Editors/YarnProjectImporterEditor.cs
@@ -35,7 +35,8 @@ namespace Yarn.Unity.Editor
         // YarnProjectImporter. Used during Inspector GUI drawing.
         internal static SerializedProperty CurrentProjectDefaultLanguageProperty;
 
-        const string ProjectUpgradeHelpURL = "https://docs.yarnspinner.dev";
+        const string ProjectUpgradeHelpURL = "https://docs.yarnspinner.dev/using-yarnspinner-with-unity/importing-yarn-files/yarn-projects#upgrading-yarn-projects";
+        const string CreateNewIssueURL = "https://github.com/YarnSpinnerTool/YarnSpinner-Unity/issues/new?assignees=&labels=bug&projects=&template=bug_report.md&title=Project Import Error";
 
         private SerializedProperty compileErrorsProperty;
         private SerializedProperty serializedDeclarationsProperty;
@@ -207,11 +208,31 @@ namespace Yarn.Unity.Editor
 
             ui.styleSheets.Add(yarnProjectStyleSheet);
 
-            if (importData == null || importData.ImportStatus == ProjectImportData.ImportStatusCode.NeedsUpgradeFromV1) {
-                ui.Add(CreateUpgradeUI(yarnProjectImporter));
+            // if the import data is null it means import has crashed
+            // we need to let the user know and perhaps ask them to file an issue
+            if (importData == null)
+            {
+                ui.Add(CreateCriticalErrorUI());
                 return ui;
             }
 
+            // next we need to handle the two edge cases
+            // either the importData is for the older format
+            // or it's completely unknown (most likely a error)
+            // in both cases we show the respective custom UI and return
+            switch (importData.ImportStatus)
+            {
+                case ProjectImportData.ImportStatusCode.NeedsUpgradeFromV1:
+                {
+                    ui.Add(CreateUpgradeUI(yarnProjectImporter));
+                    return ui;
+                }
+                case ProjectImportData.ImportStatusCode.Unknown:
+                {
+                    ui.Add(CreateUnknownErrorUI());
+                    return ui;
+                }
+            }
 
             var importDataSO = new SerializedObject(importData);
             var diagnosticsProperty = importDataSO.FindProperty(nameof(ProjectImportData.diagnostics));
@@ -510,7 +531,8 @@ namespace Yarn.Unity.Editor
             return base.HasModified() || AnyModifications;
         }
         
-        public VisualElement CreateUpgradeUI(YarnProjectImporter importer) {
+        private VisualElement CreateUpgradeUI(YarnProjectImporter importer)
+        {
             var ui = new VisualElement();
 
             var box = new VisualElement();
@@ -546,6 +568,56 @@ namespace Yarn.Unity.Editor
             ui.Add(new IMGUIContainer(ApplyRevertGUI));
 
             return ui;
+        }
+
+        private VisualElement CreateErrorUI(string headerText, string[] labels, string linkLabel, string link)
+        {
+            var ui = new VisualElement();
+            var box = new VisualElement();
+            box.AddToClassList("help-box");
+
+            Label header = new Label(headerText);
+            header.style.unityFontStyleAndWeight = FontStyle.Bold;
+            box.Add(header);
+
+            foreach (var label in labels)
+            {
+                box.Add(new Label(label));
+            }
+
+            var learnMoreLink = new Label(linkLabel);
+            learnMoreLink.RegisterCallback<MouseDownEvent>(evt =>
+            {
+                Application.OpenURL(link);
+            });
+            learnMoreLink.AddToClassList("link");
+            box.Add(learnMoreLink);
+
+            ui.Add(box);
+
+            ui.Add(new IMGUIContainer(ApplyRevertGUI));
+
+            return ui;
+        }
+
+        private VisualElement CreateCriticalErrorUI()
+        {
+            string[] labels = { 
+                "This is likely due to a bug on our end, and not in your project.",
+                "Try recreating the project and see if this resolves the issue."
+            };
+
+            return CreateErrorUI("This project has failed to import due to an internal error.", labels, "If the issue persists, please open an issue.", CreateNewIssueURL);
+        }
+
+        private VisualElement CreateUnknownErrorUI()
+        {
+            string[] labels = { 
+                "The type of this Yarn Project is unknown.",
+                "Try recreating the project and see if this resolves the issue."
+            };
+
+            return CreateErrorUI("This project has failed to import correctly.", labels, "If the issue persists, please open an issue.", CreateNewIssueURL);
         }
     }
 }